+++ examples/progress_bat_2.rs
@@ -5,6 +5,12 @@ fn main() {
     let current_dir = env::current_dir().expect("Failed to get current directory");
     let logo_path = current_dir.join("examples/progress_logo.png");
 
+    let tag = "star-rail";
+    let title = "Honkai: Star Rail";
+    let mut status = String::from("Downloading...");
+    let mut value = 0.0;
+    let mut value_string = String::from("0%");
+
     WinToastNotify::new()
         .set_duration(Duration::Long)   
         .set_title("Downloading miHoYo Game...")
@@ -12,23 +18,20 @@ fn main() {
             "May This Journey Lead Us Starward"
         ])
         .set_logo(logo_path.to_str().expect("Failed to convert path to string"), CropCircle::True)
-        .set_progress(Progress {
-            tag: "star-rail",
-            title:"Honkai: Star Rail",
-            status:"Downloading...",
-            value: 0.0,
-            value_string: "0%"
-        })
+        .set_progress(Progress {tag, title, status, value, value_string} )
         .show()
         .expect("Failed to show toast notification");
 
     for i in 1..=10 {
-        std::thread::sleep(std::time::Duration::from_secs(1));
-        let i_f32 = i as f32 / 10.0;
+        std::thread::sleep(std::time::Duration::from_millis(500));
+        value = i as f32 / 10.0;
         if i != 10 {
-            WinToastNotify::progress_update(None, "star-rail", i_f32, &format!("{:.1}%", i_f32 * 100.0)).expect("Failed to update");
+            value_string = format!("{:.1}%", value * 100.0);
+            WinToastNotify::progress_update(None, tag, value, value_string).expect("Failed to update");
         } else {
-            WinToastNotify::progress_complete(None, "star-rail", "Completed", "100%").expect("Failed to complete");
+            status = String::from("Completed");
+            value_string = String::from("100%");
+            WinToastNotify::progress_complete(None, tag, status, value_string).expect("Failed to complete");
         };
     };
 }