+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+## 0.1.4 => 0.1.5
+```rust
+// modify❗
+pub struct Action {
+    pub activation_type: ActivationType,
+    pub action_content: &'static str,
+    pub arguments: String,
+    pub image_url: Option<String>,
+}
+👇
+pub struct Action<'a> {
+    pub activation_type: ActivationType,
+    pub action_content: &'a str,
+    pub arguments: &'a str,
+    pub image_url: Option<&'a str>,
+}
+
+pub struct Progress {
+    pub tag: &'static str,
+    pub title: &'static str,
+    pub status: &'static str,
+    pub value_string: &'static str,
+}
+👇
+pub struct Progress {
+    pub tag: &'a str,
+    pub title: &'a str,
+    pub status: String,
+    pub value_string: String,
+}
+```
+```rust
+// Fix errors in examples code
+// Fix doctest code in library
+```
+
+
 ## 0.1.3 => 0.1.4
 ```rust
 // modify❗